+++ Classroom Learning Partner/Services/IDataService.cs
@@ -48,7 +48,7 @@ namespace Classroom_Learning_Partner.Services
 
         BitmapImage GetImage(string imageHashID, IPageObject pageObject);
         void LoadAllNotebookPages(Notebook notebook, bool isLoadingSubmissions = true);
-        void LoadRangeOfNotebookPages(Notebook notebook, List<decimal> pageNumbers, bool isLoadingSubmissions = true);
+        void LoadRangeOfNotebookPages(Notebook notebook, List<int> pageNumbers, bool isLoadingSubmissions = true);
 
         CacheInfo CurrentCacheInfo { get; set; }
         List<NotebookInfo> LoadedNotebooksInfo { get; }
